requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
cfgv==3.4.0
distlib==0.3.8
Django==4.2.14
django-modeltranslation==0.19.9
django-redis==5.4.0
django-resized==1.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
exceptiongroup==1.2.2
factory-boy==3.3.0
Faker==25.9.2
fakeredis==2.23.3
filelock==3.15.4
gunicorn==22.0.0
identify==2.6.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
nodeenv==1.9.1
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.6.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytest==8.2.1
pytest-django==4.8.0
pytest-factoryboy==2.7.0
pytest-json-report==1.5.0
pytest-metadata==3.1.1
pytest-mock==3.14.0
pytest-order==1.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
redis==5.0.7
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
virtualenv==20.26.3
